\name{design.mat}
\alias{design.mat}
\title{
Build a design matrix
}
\description{
Build a design matrix for ComBat adjustment.
}
\usage{
design.mat(saminfo)
}

\arguments{
  \item{saminfo}{
Matrix of sample information.
}
}

\value{
Design matrix.
}
\references{
W. Johnson E., L. Chen, Rabinovic, and A. Adjusting batch effects in microarray
expression data using Empirical Bayes methods. Biostatistics, 8(1):118-127, January2007. ISSN 1465-4644. doi: \url{http://dx.doi.org/10.1093/biostatistics/kxj037}.
}
\author{
WE Johnson
}
 \section{Warning}{This function is not called by the user directly.} 

