\name{detFileName}
\alias{detFileName}

\title{
Determine the name of a file.
}
\description{
Determine the name of a file.
}
\usage{
detFileName(file.name)
}

\arguments{
  \item{file.name}{
A character string specifying the name of the expression file to display.
}
}

\value{
\code{file.name} in upper case (in the case of a "gse" file) or without any change.
}

\author{
Haleh Yasrebi
}
 \section{Warning }{This function is not called by the user directly.} 

