\name{init.plot}
\alias{init.plot}

\title{
Start plotting
}
\description{
Plot the coordinates with the main title and axis labels.
}
\usage{
init.plot(file.name)
}

\arguments{
  \item{file.name}{
A character string specifying the name of the expression file to display.
}
}

\value{
None.
}

\author{
Haleh Yasrebi
}
 \section{Warning }{This function is not called by the user directly.} 
\keyword{Plot}
