\name{list.batch}
\alias{list.batch}

\title{
Make a list of data batches.
}
\description{
Append the batches of data sets.
}
\usage{
list.batch(saminfo)
}

\arguments{
  \item{saminfo}{
Matrix of sample information.
}
}

\value{
List of batches.
}
\references{
W. Johnson E., L. Chen, Rabinovic, and A. Adjusting batch effects in microarray
expression data using Empirical Bayes methods. Biostatistics, 8(1):118-127, January 2007. ISSN 1465-4644. doi:\url{http://dx.doi.org/10.1093/biostatistics/kxj037.}
\url{http://jlab.byu.edu/ComBat/Usage.html}

}
\author{
WE Johnson

}
\seealso{
\code{\link{ComBat}} 
}

 \section{Warning }{This function is not called by the user directly.} 
\keyword{ComBat}

