\name{trim.dat}
\alias{trim.dat}

\title{
Trim the data.
}
\description{
Trim the data of extra columns with array names 'X' or start with 'X.'
}
\usage{
trim.dat(dat)
}

\arguments{
  \item{dat}{
Matrix of gene expression data.
}
}

\value{
Matrix of gene expression data trimmed as specified in Description section.
}
\references{
W. Johnson E., L. Chen, Rabinovic, and A. Adjusting batch effects in microarray
expression data using Empirical Bayes methods. Biostatistics, 8(1):118-127, January2007. ISSN 1465-4644. doi: \url{http://dx.doi.org/10.1093/biostatistics/kxj037}.

}
\author{
WE Johnson
}
 \section{Warning }{This function is not called by the user directly.} 
